Refatorando Tratamento de Erros Personalizado + Error 404 NotFound

+++ src/controllers/livrosController.js
@@ -2,6 +2,9 @@ import livro from "../models/Livro.js";
 import { autor } from "../models/Autor.js";   //Ao importar de scripts que não tem uma exportação DEFAULT, importar como se fosse um objeto, e extraindo somente a parte necessaria dele
 //import autor from "../models/Autor.js";     //Se importar assim, daria erro
 
+import ErroValidacao from "../erros/ErroValidacao.js";
+import mongoose from "mongoose";
+
 class LivroController {
 
   //Em conexões com o banco de dados é sempre interessante utilziar metodos asyncronos para não alterar mesmo registro ao mesmo tempo.
@@ -21,8 +24,11 @@ class LivroController {
 
   static listarLivros = async (req, res, next) => {   
     try {      
-      //const listaLivros = await livro.find({});       //recebendo autor por embedding
 
+      //Testando o Manipulador de Erros.:
+      //throw new ErroValidacao() ; //throw new NaoEncontrado(); // mongoose.Error.ValidationError ;//mongoose.Error.CastError(); //ErroBase(); //Error();
+
+      //const listaLivros = await livro.find({});       //Se fosse receber autor por embedding
       //Recebendo por Referencia ao invés de embbeding
       //Quando usamos references o autor não faz mais parte do objeto livro. Assim, cada livro deve ser “populado” com as referências do autor.
       //.populate("autor").exec()
@@ -38,7 +44,13 @@ class LivroController {
     try {
       const id = req.params.id;
       const livroEncontrado = await livro.findById(id);
-      res.status(200).json(livroEncontrado);
+
+      if(livroEncontrado){
+        res.status(200).json(livroEncontrado);
+      }else{
+        next(new NaoEncontrado("Livro não encontrado"));
+      } 
+
     } catch (erro) {
       next( erro ) ; 
       //res.status(500).json({ message: `${erro.message} - falha na requisição do livro` });
@@ -79,8 +91,15 @@ class LivroController {
   static atualizarLivro = async (req, res, next) => {
     try {
       const id = req.params.id;
-      await livro.findByIdAndUpdate(id, req.body);
-      res.status(200).json({ message: "livro atualizado" });
+
+      const livroEncontrado = await livro.findById(id);
+      if(livroEncontrado){
+        await livro.findByIdAndUpdate(id, req.body);  
+        res.status(200).json({ message: "livro atualizado" });
+      }else{
+        next(new NaoEncontrado("Livro não encontrado"));
+      } 
+      
     } catch (erro) {
       next( erro ) ; 
       //res.status(500).json({ message: `${erro.message} - falha na atualização` });
@@ -90,8 +109,15 @@ class LivroController {
   static excluirLivro = async (req, res, next) => {
     try {
       const id = req.params.id;
-      await livro.findByIdAndDelete(id);
-      res.status(200).json({ message: "livro excluído com sucesso" });
+
+      const livroEncontrado = await livro.findById(id);
+      if(livroEncontrado){
+        await livro.findByIdAndDelete(id);
+        res.status(200).json({ message: "livro excluído com sucesso" });
+      }else{
+        next(new NaoEncontrado("Livro não encontrado"));
+      } 
+      
     } catch (erro) {
       next( erro ) ; 
       //res.status(500).json({ message: `${erro.message} - falha na exclusão` });
